Fix a monkey bug - a null pointer issue in native.
A monkey report shows that we try to generate a mosaic even before we
added any frame to the aligner. This is unlikely to be triggered by human
actions but machine generated sequences. To fix it, we ensure that we do
blending on frames only after we have accepted frames in aligner.

Bug: 5182097

Change-Id: I16c8efea2b1e3f5b53e65e480da169e24c939ec2